pyproject.toml
[project]
name = "youtube-resume"
version = "0.1.0"
description = "Resume your Youtbe video content"
authors = [
    {name = "Johan Beaucé",email = "johan.beauce@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "youtube-transcript-api (>=1.1.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.90.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
